backend/requirements.txt
Flask==0.10.1
Flask-API==0.6.3
Flask-JWT==0.3.2
Flask-Login==0.2.11
Flask-Migrate==1.8.0
Flask-OAuth==0.12
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
numpy==1.11.2
PyJWT==1.4.0
SQLAlchemy==1.0.14
WTForms==2.0.2
Werkzeug==0.10.1
alembic==0.8.4
argparse==1.2.1
bcrypt==2.0.0
boto3==1.4.1
botocore==1.4.78
flask-restful-swagger==0.19
httplib2==0.9
itsdangerous==0.24
oauth2==1.5.211
requests==2.6.0
rfc3339==5
structlog==16.1.0
twilio==5.3.0
ujson==1.35
uwsgidecorators==1.1.0
wsgiref==0.1.2
